move components API to a plugin, add composeWithTracker to it
