+++ imports/plugins/core/components/client/index.js
@@ -1,0 +1,1 @@
+export * from "../lib";
