+++ imports/plugins/core/components/package.json
@@ -1,0 +1,12 @@
+{
+  "name": "@reactioncommerce/reaction-components",
+  "version": "1.0.0",
+  "description": "",
+  "main": "client/index.js",
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1"
+  },
+  "keywords": [],
+  "author": "",
+  "license": "ISC"
+}
