+++ imports/plugins/core/dashboard/client/containers/actionViewContainer.js
@@ -1,7 +1,7 @@
 import React from "react";
 import { StyleRoot } from "radium";
 import _ from "lodash";
-import { composeWithTracker } from "/lib/api/compose";
+import { composeWithTracker } from "@reactioncommerce/reaction-components";
 import { Reaction } from "/client/api";
 import { TranslationProvider, AdminContextProvider } from "/imports/plugins/core/ui/client/providers";
