+++ imports/plugins/core/email/client/containers/emailSettings.js
@@ -1,10 +1,9 @@
 import { compose, withProps } from "recompose";
-import { registerComponent } from "@reactioncommerce/reaction-components";
+import { registerComponent, composeWithTracker } from "@reactioncommerce/reaction-components";
 import { Meteor } from "meteor/meteor";
 import { Reaction } from "/client/api";
 import { saveSettings } from "../actions";
 import EmailSettings from "../components/emailSettings";
-import { composeWithTracker } from "/lib/api/compose";
 
 const providers = Object.keys(require("nodemailer-wellknown/services.json"));
