+++ imports/plugins/core/layout/client/containers/adminContainer.js
@@ -1,4 +1,4 @@
-import { composeWithTracker } from "/lib/api/compose";
+import { composeWithTracker } from "@reactioncommerce/reaction-components";
 import { Reaction } from "/client/api";
 
 function composer(props, onData) {
