+++ imports/plugins/core/orders/client/containers/ordersListContainer.js
@@ -1,7 +1,7 @@
 import React, { Component } from "react";
 import PropTypes from "prop-types";
+import { composeWithTracker } from "@reactioncommerce/reaction-components";
 import { Meteor } from "meteor/meteor";
-import { composeWithTracker } from "/lib/api/compose";
 import { Media } from "/lib/collections";
 import { Reaction } from "/client/api";
 import OrdersList from "../components/ordersList.js";
