+++ imports/plugins/core/orders/client/containers/totalActionsContainer.js
@@ -1,6 +1,6 @@
 import React, { Component } from "react";
 import PropTypes from "prop-types";
-import { composeWithTracker } from "/lib/api/compose";
+import { composeWithTracker } from "@reactioncommerce/reaction-components";
 import TotalActions from "../components/totalActions";
 
 class TotalActionsContaner extends Component {
