+++ imports/plugins/core/revisions/client/containers/settingsContainer.js
@@ -1,6 +1,6 @@
 import React, { Component } from "react";
 import PropTypes from "prop-types";
-import { composeWithTracker } from "/lib/api/compose";
+import { composeWithTracker } from "@reactioncommerce/reaction-components";
 import SettingsComponent from "../components/settings";
 import { Packages } from "/lib/collections";
 import { Meteor } from "meteor/meteor";
