+++ imports/plugins/core/ui-navbar/client/containers/navbar.js
@@ -1,5 +1,4 @@
-import { registerComponent } from "@reactioncommerce/reaction-components";
-import { composeWithTracker } from "/lib/api/compose";
+import { registerComponent, composeWithTracker } from "@reactioncommerce/reaction-components";
 import { Reaction } from "/client/api";
 import NavBar from "../components/navbar";
