+++ imports/plugins/core/ui-tagnav/client/containers/tagNavContainer.js
@@ -2,9 +2,8 @@ import _ from "lodash";
 import update from "react/lib/update";
 import React, { Component } from "react";
 import PropTypes from "prop-types";
-import { Components, registerComponent } from "@reactioncommerce/reaction-components";
+import { Components, registerComponent, composeWithTracker } from "@reactioncommerce/reaction-components";
 import { Reaction, Router } from "/client/api";
-import { composeWithTracker } from "/lib/api/compose";
 import { getTagIds } from "/lib/selectors/tags";
 import { TagHelpers } from "/imports/plugins/core/ui-tagnav/client/helpers";
 import { Tags } from "/lib/collections";
