+++ imports/plugins/core/ui/client/containers/edit.js
@@ -1,8 +1,7 @@
 import React, { Children, Component } from "react";
 import PropTypes from "prop-types";
-import { Components, registerComponent } from "@reactioncommerce/reaction-components";
+import { Components, registerComponent, composeWithTracker } from "@reactioncommerce/reaction-components";
 import { Reaction } from "/client/api";
-import { composeWithTracker } from "/lib/api/compose";
 
 const styles = {
   editContainerItem: {
