+++ imports/plugins/core/ui/client/containers/tagListContainer.js
@@ -2,10 +2,9 @@ import React, { PureComponent } from "react";
 import PropTypes from "prop-types";
 import _ from "lodash";
 import update from "react/lib/update";
-import { registerComponent } from "@reactioncommerce/reaction-components";
+import { registerComponent, composeWithTracker } from "@reactioncommerce/reaction-components";
 import { Meteor } from "meteor/meteor";
 import { Reaction, i18next } from "/client/api";
-import { composeWithTracker } from "/lib/api/compose";
 import { TagList } from "../components";
 import { Tags } from "/lib/collections";
 import { getTagIds } from "/lib/selectors/tags";
