+++ imports/plugins/core/ui/client/providers/adminContextProvider.js
@@ -1,6 +1,6 @@
 import React, { Component, Children } from "react"; // eslint-disable-line
 import PropTypes from "prop-types";
-import { composeWithTracker } from "/lib/api/compose";
+import { composeWithTracker } from "@reactioncommerce/reaction-components";
 
 class AdminContextProvider extends Component {
   static childContextTypes = {
