+++ imports/plugins/included/product-variant/containers/productsContainer.js
@@ -1,12 +1,11 @@
 import React, { PureComponent } from "react";
 import PropTypes from "prop-types";
-import { registerComponent } from "@reactioncommerce/reaction-components";
+import { registerComponent, composeWithTracker } from "@reactioncommerce/reaction-components";
 import { Meteor } from "meteor/meteor";
 import { Session } from "meteor/session";
 import { Reaction } from "/client/api";
 import { ITEMS_INCREMENT } from "/client/config/defaults";
 import { ReactionProduct } from "/lib/api";
-import { composeWithTracker } from "/lib/api/compose";
 import { applyProductRevision } from "/lib/api/products";
 import { Products, Tags } from "/lib/collections";
 import ProductsComponent from "../components/products";
