+++ imports/plugins/included/social/client/containers/socialSettingsContainer.js
@@ -1,8 +1,8 @@
 import React, { Component } from "react";
 import PropTypes from "prop-types";
 import { isEqual } from "lodash";
+import { composeWithTracker } from "@reactioncommerce/reaction-components";
 import { Meteor } from "meteor/meteor";
-import { composeWithTracker } from "/lib/api/compose";
 import { Reaction, i18next } from "/client/api";
 import { Packages } from "/lib/collections";
 import { SocialSettings } from "../components";
