+++ lib/api/index.js
@@ -1,7 +1,5 @@
 export * from "./account-validation";
 export { default as Catalog } from "./catalog";
-export * from "./components";
-export * from "./compose";
 export { default as ReactionProduct } from "./products";
 export { PropTypes } from "./prop-types";
 export * from "./helpers";
